update path test
