+++ plugins/woocommerce/tests/php/includes/class-woocommerce-test.php
@@ -6,6 +6,24 @@ use Automattic\WooCommerce\Internal\Utilities\LegacyRestApiStub;
  * Unit tests for the WooCommerce class.
  */
 class WooCommerce_Test extends \WC_Unit_Test_Case {
+
+	/**
+	 * The default URI.
+	 *
+	 * @var string
+	 */
+	private static $default_uri;
+
+	/**
+	 * Store the default URI.
+	 *
+	 */
+	public static function setUpBeforeClass(): void {
+		parent::setUpBeforeClass();
+		self::$default_uri = isset( $_SERVER['REQUEST_URI'] ) ? $_SERVER['REQUEST_URI'] : ''; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.MissingUnslash, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+	}
+
+
 	/**
 	 * Setup test data. Called before every test.
 	 */
@@ -20,11 +38,36 @@ class WooCommerce_Test extends \WC_Unit_Test_Case {
 		wp_set_current_user( $this->user );
 	}
 
+
+	/**
+	 * Restore the default URI.
+	 */
+	public static function tearDownAfterClass(): void {
+		parent::tearDownAfterClass();
+		$_SERVER['REQUEST_URI'] = self::$default_uri;
+	}
+
 	/**
 	 * Test that the $api property is defined and holds an instance of LegacyRestApiStub
 	 * (the Legacy REST API was removed in WooCommerce 9.0).
 	 */
 	public function test_api_property(): void {
 		$this->assertInstanceOf( LegacyRestApiStub::class, WC()->api );
+	}
+
+	/**
+	 * Test that the rest api returns false when it is not an rest api request.
+	 */
+	public function test_rest_api_returns_false() {
+		$this->assertEquals( WC()->is_rest_api_request(), false );
+	}
+
+	/**
+	 * Test that the rest api returns true when it is an rest api request.
+	 */
+	public function test_rest_api_returns_true() {
+		// Set the request uri to a rest api request.
+		$_SERVER['REQUEST_URI'] = '/wp-json/wc/v3/products';
+		$this->assertEquals( WC()->is_rest_api_request(), true );
 	}
 }
