--- plugins/woocommerce/tests/php/src/Internal/RestApiTest.php
@@ -1,50 +1,0 @@
-<?php
-
-namespace Automattic\WooCommerce\Tests\Internal;
-
-/**
- * Tests for the RestApiTest class.
- */
-class RestApiTest extends \WC_Unit_Test_Case {
-
-	/**
-	 * The default URI.
-	 *
-	 * @var string
-	 */
-	private static $default_uri;
-
-	/**
-	 * Store the default URI.
-	 *
-	 */
-	public static function setUpBeforeClass(): void {
-		parent::setUpBeforeClass();
-		self::$default_uri = isset( $_SERVER['REQUEST_URI'] ) ? $_SERVER['REQUEST_URI'] : ''; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.MissingUnslash, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
-	}
-
-	/**
-	 * Restore the default URI.
-	 */
-	public static function tearDownAfterClass(): void {
-		parent::tearDownAfterClass();
-		$_SERVER['REQUEST_URI'] = self::$default_uri;
-	}
-
-
-	/**
-	 * Test that the rest api returns false when it is not an rest api request.
-	 */
-	public function test_rest_api_returns_false() {
-		$this->assertEquals( WC()->is_rest_api_request(), false );
-	}
-
-	/**
-	 * Test that the rest api returns true when it is an rest api request.
-	 */
-	public function test_rest_api_returns_true() {
-		// Set the request uri to a rest api request.
-		$_SERVER['REQUEST_URI'] = '/wp-json/wc/v3/products';
-		$this->assertEquals( WC()->is_rest_api_request(), true );
-	}
-}
